requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.10.0
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.58.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.20
langchain-core==0.3.63
langchain-openai==0.3.11
langchain-text-splitters==0.3.8
langsmith==0.3.19
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
mysql-connector-python==9.3.0
numpy==2.2.4
openai==1.68.2
orjson==3.10.16
packaging==24.2
pdfkit==1.0.0
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydyf==0.11.0
pypdf==5.4.0
pyphen==0.17.2
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.1
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
yarl==1.18.3
zopfli==0.2.3.post1
zstandard==0.23.0
